create moment helper functions for availability table and postings details
modified:   frontend/package.json
	modified:   frontend/src/components/common/PostingDetails.tsx
		modified:   frontend/src/components/pages/volunteer/posting/VolunteerPostingsPage.tsx
	modified:   frontend/src/components/volunteer/PostingCard.tsx
	modified:   frontend/src/components/volunteer/shifts/VolunteerAvailabilityTableRow.tsx
	new file:   frontend/src/utils/DateTimeUtils.ts
	deleted:    frontend/src/utils/DateUtils.ts
	modified:   frontend/src/utils/TimeUtil.ts
	modified:   frontend/yarn.lock
