+++ frontend/package.json
@@ -35,6 +35,7 @@
     "json-schema": "^0.4.0",
     "json2csv": "^5.0.6",
     "jsonwebtoken": "^8.5.1",
+    "moment": "^2.29.1",
     "react": "^17.0.1",
     "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.1",
