+++ frontend/src/components/volunteer/PostingCard.tsx
@@ -11,8 +11,8 @@ import {
   Tag,
 } from "@chakra-ui/react";
 
-import { formatDateStringWithYear } from "../../utils/DateUtils";
 import { SkillResponseDTO } from "../../types/api/SkillTypes";
+import { formatDateStringYear } from "../../utils/DateTimeUtils";
 
 type PostingCardProps = {
   id: string;
@@ -54,9 +54,9 @@ const PostingCard = ({
           </Text>
           <Text textStyle="caption">
             <CalendarIcon w={4} pr={1} />
-            {`${formatDateStringWithYear(
-              startDate,
-            )} - ${formatDateStringWithYear(endDate)}`}
+            {`${formatDateStringYear(startDate)} - ${formatDateStringYear(
+              endDate,
+            )}`}
           </Text>
         </HStack>
         <Text textStyle="body-regular" noOfLines={2}>
@@ -75,7 +75,7 @@ const PostingCard = ({
         </Box>
         <HStack justifyContent="space-between" w="100%">
           <Text textStyle="caption" color="text.gray">
-            Deadline: {formatDateStringWithYear(autoClosingDate)}
+            Deadline: {formatDateStringYear(autoClosingDate)}
           </Text>
           <ButtonGroup spacing="10px" size="md">
             <Button variant="ghost">View Details</Button>
